turn DevToolsFactory into an interface for mockability
